LayoutTests/fast/forms/month-multiple-fields/month-multiple-fields-preserve-value-after-history-back.html

<!DOCTYPE html>
<html>
<head>
<script src="../../../fast/js/resources/js-test-pre.js"></script>
</head>
<body>
<iframe id="iframe"></iframe>
<script>
description('Checks input value is preserved after backward and forward');
if (!window.eventSender || !window.internals)
    debug('Please run within DRT/WTR');
window.jsTestIsAsync = true;

var state = 'direct';
var testCases = [];

function failed(message) {
    testFailed(message + ' in ' + state);
    finishJSTest();
}

function finishTest() {
    runTestCases();
    finishJSTest();
}

function runTestCases() {
    for (var index = 0; index < testCases.length; ++index) {
        var testCase = testCases[index];
        debug(testCase['description'] + ':');
        shouldBeEqualToString('"' + testCase['actual'] + '"', testCase['expected']);
    }
    debug('');
}

window.onload = function() {
    document.getElementById('iframe').src = 'resources/preserve-value-after-history-back-frame.html?page1';
};
</script>
<script src="../../../fast/js/resources/js-test-post.js"></script>
</body>
</html>
